move code converter logic into separate file
